+++ cmd/gghstats/serve.go
@@ -254,7 +254,9 @@ func runServe(args []string) error {
 		defer rateLimiter.Shutdown()
 	}
 
+	trusted := server.ParseTrustedProxies(os.Getenv("GGHSTATS_TRUSTED_PROXIES"))
 	whitelist := server.NewWhitelist(server.ParseWhitelistEnv(), cfg.APIToken)
+	server.WarnTrustedProxiesIfNeeded(trusted, rateLimiter != nil, whitelist != nil)
 
 	ctx, stop := signal.NotifyContext(context.Background(), syscall.SIGINT, syscall.SIGTERM)
 	defer stop()
@@ -323,6 +325,7 @@ func runServe(args []string) error {
 		DefaultLocale:     i18n.EnvDefaultLocale(),
 		EnabledLocales:    i18n.EnvEnabledLocales(),
 		RateLimiter:       rateLimiter,
+		TrustedProxies:    trusted,
 		Whitelist:         whitelist,
 		HeadHTML:          template.HTML(cfg.HeadHTML),
 		ReverseProxyRules: server.ParseReverseProxyRules(cfg.ReverseProxyRules),
@@ -332,8 +335,12 @@ func runServe(args []string) error {
 
 	addr := cfg.Host + ":" + cfg.Port
 	srv := &http.Server{
-		Addr:    addr,
-		Handler: handler,
+		Addr:              addr,
+		Handler:           handler,
+		ReadHeaderTimeout: 10 * time.Second,
+		ReadTimeout:       30 * time.Second,
+		WriteTimeout:      60 * time.Second,
+		IdleTimeout:       120 * time.Second,
 	}
 
 	return serveHTTP(ctx, srv, cfg, cancel)